Feature/party - 파티 API 에러 정립 및 문서화/파티 유저 포지션 변경추가

+++ src/party/application/command/delete-party.handler.ts
@@ -1,4 +1,4 @@
-import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
 import { PartyFactory } from 'src/party/domain/party/party.factory';
@@ -17,10 +17,17 @@ export class DeletePartyHandler implements ICommandHandler<DeletePartyCommand> {
 
   async execute(command: DeletePartyCommand) {
     const { userId, partyId } = command;
+
+    const findParty = await this.partyRepository.findOne(partyId);
+
+    if (findParty) {
+      throw new NotFoundException('PARTY_NOT_EXIST', '파티를 찾을 수 없습니다.');
+    }
+
     const partyUser = await this.partyUserRepository.findOne(userId, partyId);
 
     if (partyUser.authority !== 'master') {
-      throw new ForbiddenException('권한이 없습니다.');
+      throw new ForbiddenException('ACCESS_DENIED', '파티 삭제 권한이 없습니다.');
     }
 
     await this.partyRepository.deleteById(partyId);
